refactor: move recording logic to dedicated class
